+++ src/components/Schematics/Schematic-Flow.vue
@@ -1,9 +1,9 @@
 <script>
 import { mapGetters } from 'vuex'
 import * as d3_base from 'd3'
-import { event } from 'd3-selection'
 import { zoom } from 'd3-zoom'
 import uniqueId from 'lodash.uniqueid'
+import throttle from 'lodash.throttle'
 import Legend from '@/components/Schematics/Legend'
 import LogRocket from 'logrocket'
 import SchematicNode from '@/components/Schematics/Schematic-Node'
@@ -15,7 +15,7 @@ import {
   isDiagonal
 } from '@/utils/curveMetro'
 import { STATE_COLORS } from '@/utils/states'
-import SchematicWorker from 'workerize-loader?inline!./SchematicWorker'
+import SchematicWorker from 'workerize-loader?inline!@/workers/schematic'
 
 // Add our custom curve to d3 base
 // and then merges the libraries together into one
@@ -89,6 +89,7 @@ export default {
       showLabels: false,
       showTooltip: true,
       size: 0,
+      timer: null,
       tooltipData: null,
       transitionDuration: 250,
       transform: null,
@@ -136,6 +137,9 @@ export default {
         level2: this.transform?.k / this.size > 0.15
       }
     },
+    animateCanvas: function() {
+      return throttle(this.rawAnimateCanvas, 16)
+    },
     visibleNodes() {
       if (!this.transform) return []
       const buf = 100 // buffer
@@ -161,13 +165,7 @@ export default {
       }
     },
     showCards() {
-      let t = d3.timer(elapsed => {
-        this.drawCanvas(this.transform)
-
-        if (elapsed > this.transitionDuration) {
-          t.stop()
-        }
-      })
+      this.animateCanvas()
     }
   },
   mounted() {
@@ -180,33 +178,33 @@ export default {
     }
   },
   beforeDestroy() {
-    this.id = uniqueId('schematic')
+    cancelAnimationFrame(this.drawCanvas)
+    this.id = null
     this.hasFit = null
     this.nodeData = null
     this.worker = null
+    this.custom = null
+    this.timer.stop()
+    this.timer = null
   },
   methods: {
-    _zoomed() {
-      let zoomEvent = event
-
+    _zoomed(event) {
       // Disabling for now to prevent flickering when throttling
       // drawing the canvas
       // this.context.clearRect(0, 0, this.width, this.height)
 
-      if (zoomEvent && zoomEvent.sourceEvent && zoomEvent.sourceEvent.ctrlKey) {
-        if (zoomEvent.sourceEvent.deltaY > 0) {
-          zoomEvent.transform.k =
-            zoomEvent.transform.k - zoomEvent.transform.k * 0.03
-        } else if (zoomEvent && zoomEvent.sourceEvent.deltaY < 0) {
-          zoomEvent.transform.k =
-            zoomEvent.transform.k + zoomEvent.transform.k * 0.03
+      if (event && event.sourceEvent && event.sourceEvent.ctrlKey) {
+        if (event.sourceEvent.deltaY > 0) {
+          event.transform.k = event.transform.k - event.transform.k * 0.03
+        } else if (event && event.sourceEvent.deltaY < 0) {
+          event.transform.k = event.transform.k + event.transform.k * 0.03
         }
       }
-      this.transform = zoomEvent.transform
-      this.transformEventK = zoomEvent.transform.k
-      this.transformEventX = zoomEvent.transform.x
-      this.transformEventY = zoomEvent.transform.y
-      this.drawCanvas(this.transform)
+      this.transform = event.transform
+      this.transformEventK = event.transform.k
+      this.transformEventX = event.transform.x
+      this.transformEventY = event.transform.y
+      requestAnimationFrame(this.drawCanvas)
     },
     _zoomIn() {
       if (this.transform.k < 0.01) return
@@ -267,7 +265,7 @@ export default {
       this.loadingKey--
     },
     async createDag(tasks) {
-      if (tasks.length === 0) return
+      if (tasks.length === 0 || this.loading) return
       this.loadingKey += 2
       this.mappedTasks = {}
 
@@ -414,19 +412,52 @@ export default {
       this.redraw()
     },
     fitViz() {
-      this.canvas
-        .transition()
-        .duration(500)
-        .call(
+      this.canvas.call(
+        this.scaledZoom.transform,
+        d3.zoomIdentity
+          .translate(this.canvasAdjustment.w / 2, this.canvasAdjustment.h / 2)
+          .scale(1)
+      )
+    },
+    rawAnimateCanvas() {
+      cancelAnimationFrame(this.drawCanvas)
+
+      this.timer?.stop()
+      this.timer = d3.timer(elapsed => {
+        this.canvas.call(
           this.scaledZoom.transform,
           d3.zoomIdentity
-            .translate(this.canvasAdjustment.w / 2, this.canvasAdjustment.h / 2)
-            .scale(1)
+            .translate(this.transformEventX, this.transformEventY)
+            .scale(this.transformEventK)
         )
+
+        if (elapsed > this.transitionDuration + 100) {
+          setTimeout(() => {
+            // Final fit
+            this.painting = false
+
+            // Fit after the first drawing
+            if (!this.hasFit) {
+              this.hasFit = true
+              this.fitViz()
+              setTimeout(() => {
+                this.showCanvas = true
+              }, 600)
+            }
+          }, 500)
+
+          this.timer.stop()
+        }
+      })
     },
-    drawCanvas(transform) {
+    drawCanvas() {
+      // If the custom ref has been destroyed we don't want to continue trying
+      // to draw the canvas (this can happen if this has been queued)
+      if (!this.custom) return
+
       this.context.clearRect(0, 0, this.width, this.height)
 
+      const transform = this.transform
       const context = this.context
       const showNodes = !this.showCards
       const showDetails = this.showDetails
@@ -442,6 +473,8 @@ export default {
         scaledArrow > 10 ? 10 : scaledArrow < 5 ? 5 : scaledArrow
       const nodeSize =
         scaledNode + Math.min(scaledNode, scaledNode * (size / scaledNode))
+
+      if (!this.custom) return
 
       let edges = this.custom.selectAll('path')
 
@@ -500,9 +533,11 @@ export default {
         })
       })
 
-      if (!showNodes && showDetails.level1) return
+      if (!showNodes && showDetails?.level1) return
 
       let nodes = this.custom.selectAll('circle')
+
+      if (!nodes) return
 
       // Resets the alpha before we draw nodes,
       // which we always want to be opaque
@@ -550,35 +585,7 @@ export default {
     redraw() {
       this.updateEdges()
       this.updateNodes()
-
-      let t = d3.timer(elapsed => {
-        this.drawCanvas(d3.zoomIdentity)
-
-        this.canvas.call(
-          this.scaledZoom.transform,
-          d3.zoomIdentity
-            .translate(this.transformEventX, this.transformEventY)
-            .scale(this.transformEventK)
-        )
-
-        if (elapsed > this.transitionDuration + 100) {
-          setTimeout(() => {
-            // Final fit
-            this.painting = false
-
-            // Fit after the first drawing
-            if (!this.hasFit) {
-              this.hasFit = true
-              this.fitViz()
-              setTimeout(() => {
-                this.showCanvas = true
-              }, 600)
-            }
-          }, 500)
-
-          t.stop()
-        }
-      })
+      this.animateCanvas()
     },
     updateNodes() {
       // Preserves references to this for event handlers
@@ -618,9 +625,6 @@ export default {
           update =>
             update
               .attr('fill', this.calcNodeColor)
-              .transition()
-              .duration(this.transitionDuration)
-              .delay((d, i) => i * 10)
               .attr('r', d => (d.id == this.selectedTaskId ? size : size * 0.9))
               .attr('opacity', 1)
               .attr('cy', data =>
@@ -689,33 +693,28 @@ export default {
                   })
               }),
           update =>
-            update
-              .transition()
-              .duration(this.transitionDuration)
-              .delay((d, i) => i * 10)
-              .attr('stroke', this.calcStrokeColor)
-              .call(update => {
-                update
-                  .transition()
-                  .duration(this.transitionDuration)
-                  .delay((d, i) => i * 10)
-                  .attr('stroke-opacity', d => {
-                    let opacity
-                    if (this.selectedTaskId) {
-                      if (
-                        d.source.id == this.selectedTaskId ||
-                        d.target.id == this.selectedTaskId
-                      ) {
-                        opacity = 1
-                      } else {
-                        opacity = 0.2
-                      }
-                    } else {
+            update.attr('stroke', this.calcStrokeColor).call(update => {
+              update
+                .transition()
+                .duration(this.transitionDuration)
+                .delay((d, i) => i * 10)
+                .attr('stroke-opacity', d => {
+                  let opacity
+                  if (this.selectedTaskId) {
+                    if (
+                      d.source.id == this.selectedTaskId ||
+                      d.target.id == this.selectedTaskId
+                    ) {
                       opacity = 1
+                    } else {
+                      opacity = 0.2
                     }
-                    return opacity
-                  })
-              }),
+                  } else {
+                    opacity = 1
+                  }
+                  return opacity
+                })
+            }),
           exit =>
             exit.call(exit =>
               exit
@@ -799,13 +798,8 @@ export default {
           query: { ...this.$route.query, schematic: '' }
         })
       }
-      let t = d3.timer(elapsed => {
-        this.drawCanvas(this.transform)
 
-        if (elapsed > this.transitionDuration + 100) {
-          t.stop()
-        }
-      })
+      this.animateCanvas()
     },
     showAllNodes() {
       this.custom
@@ -1046,7 +1040,7 @@ export default {
     >
       <v-tooltip top>
         <template v-slot:activator="{ on }">
-          <v-btn icon tile v-on="on" @click="fitViz()">
+          <v-btn icon tile v-on="on" @click="fitViz">
             <v-icon>center_focus_strong</v-icon>
           </v-btn>
         </template>