+++ src/pages/Flow/Timeline-Tile.vue
@@ -17,7 +17,8 @@ export default {
     flow: {
       required: true,
       type: Object
-    }
+    },
+    visible: { type: Boolean, default: () => true }
   },
   data() {
     return {
@@ -102,6 +103,7 @@ export default {
       :height="100"
       :min-bands="100"
       show-controls
+      :visible="visible"
       y-field="duration"
       @bar-click="_barClick"
       @bar-mouseout="_barMouseout"