+++ Changelog.md
@@ -10,6 +10,148 @@
 
 - None
 
+## 2020-09-17
+
+### Features and Improvements
+
+- Add unit tests for tenant store [#195](https://github.com/PrefectHQ/ui/pull/195)
+
+### Bugfixes
+
+- Fix a race condition with flow concurrency usage - [#231](https://github.com/PrefectHQ/ui/pull/231)
+
+## 2020-09-15
+
+### Features and Improvements
+
+- Fix app-wide component teardown by manually nulling references that were preventing the garbage colector from running correctly - [#212](https://github.com/PrefectHQ/ui/pull/212)
+- Remove `task_run` join from the Dashboard-level Upcoming Runs query - [#218](https://github.com/PrefectHQ/ui/pull/218)
+- Introduce a notifications tile - [#220](https://github.com/PrefectHQ/ui/pull/220)
+
+### Bugfixes
+
+- Fix BarChart component resize event listener removal - [#212](https://github.com/PrefectHQ/ui/pull/212)
+- Remove all BarChart underscore method and property references - [#212](https://github.com/PrefectHQ/ui/pull/212)
+- Fix visual bugs associated with resizing while the BarChart component wasn't visible - [#216](https://github.com/PrefectHQ/ui/pull/216)
+- Fix bug preventing late runs from being cleared in the UI - [#227](https://github.com/PrefectHQ/ui/pull/227)
+
+## 2020-09-08
+
+### Features and Improvements
+
+- Add unit test for the auth0 Vuex store[#162](https://github.com/PrefectHQ/ui/pull/162)
+- Remove deprecated route calls from Task Concurrency page - [#193](https://github.com/PrefectHQ/ui/pull/193)
+- Add a loading state for accepting invitations in the sidenav and a warning before a user removes themself from a tenant [#190](https://github.com/PrefectHQ/ui/pull/190)
+- Changes to the barchart visualization to take browser visibility into account when performing repaints of the canvas and svg elements - [201](https://github.com/PrefectHQ/ui/pull/201)
+- Removes `task_run` joins on the upcoming runs queries for both the dashboard and flow pages - [201](https://github.com/PrefectHQ/ui/pull/201)
+- Create service worker scaffolding - [201](https://github.com/PrefectHQ/ui/pull/201)
+- Use native JavaScript Date objects instead of `moment` for duration calculations - [201](https://github.com/PrefectHQ/ui/pull/201)
+- Export the Apollo cache - [201](https://github.com/PrefectHQ/ui/pull/201)
+- Display the token name instead of the token id on the agent card - [#200](https://https://github.com/PrefectHQ/ui/pull/200)
+- Add new Flow Concurrency page for managing user-driven flow concurrency limits - [#199](https://github.com/PrefectHQ/ui/pull/199)
+
+### Bugfixes
+
+- Keep the approve button in a loading state until the task state updates from paused to resume [#192](https://github.com/PrefectHQ/ui/pull/192)
+
+## 2020-09-1
+
+### Features and Improvements
+
+- Add unit tests for the api Vuex store [#127](https://github.com/PrefectHQ/ui/pull/127)
+- Rename "Mark As" to "Set State" for clarity - [#183](https://github.com/PrefectHQ/ui/pull/183)
+- Add link to task run page from gantt chart - [#185](https://github.com/PrefectHQ/ui/pull/185)
+- Add unit tests for the license Vuex store [#164](https://github.com/PrefectHQ/ui/pull/164)
+- Show (sorted) default parameters in the flow details tile and add an info card with link to the flow settings tab [#176](https://github.com/PrefectHQ/ui/pull/176)
+- Clean up the flow run restart dialog - [#186](https://github.com/PrefectHQ/ui/pull/186)
+
+### Bugfixes
+
+- Fix bug with Restart from Failed rerunning successfull mapped child tasks - [#182](https://github.com/PrefectHQ/ui/issues/182)
+
+## 2020-08-31a
+
+### Features and Improvements
+
+- Display number of expected mapped runs when the data exists - [#173](https://github.com/PrefectHQ/ui/issues/173)
+- August package updates [#180](https://github.com/PrefectHQ/ui/pull/180):
+  - [@babel/runtime@7.11.2](https://www.npmjs.com/package/@babel/runtime/v/7.11.2)
+  - [@vue/cli-plugin-babel@4.5.4](https://www.npmjs.com/package/@vue/cli-plugin-babel/v/4.5.4)
+  - [@vue/cli-service@4.5.4](https://www.npmjs.com/package/@vue/cli-service/v/4.5.4)
+  - [@babel/plugin-transform-runtime@7.11.5](https://www.npmjs.com/package/@babel/plugin-transform-runtime/v/7.11.5)
+  - [@vue/cli-plugin-eslint@4.5.4](https://www.npmjs.com/package/@vue/cli-plugin-eslint/v/4.5.4)
+  - [@vue/cli-plugin-unit-jest@4.5.4](https://www.npmjs.com/package/@vue/cli-plugin-unit-jest/v/4.5.4)
+  - [babel-jest@26.3.0](https://www.npmjs.com/package/babel-jest/v/26.3.0)
+  - [@auth0/auth0-spa-js@1.11.0](https://www.npmjs.com/package/@auth0/auth0-spa-js/v/1.11.0)
+  - [@babel/core@7.11.5](https://www.npmjs.com/package/@babel/core/v/7.11.5)
+  - [@vue/test-utils@1.0.5](https://www.npmjs.com/package/@vue/test-utils/v/1.0.5)
+  - [cronstrue@1.100.0](https://www.npmjs.com/package/cronstrue/v/1.100.0)
+  - [codemirror-graphql@0.12.1](https://www.npmjs.com/package/codemirror-graphql/v/0.12.1)
+  - [d3-selection@2.0.0](https://www.npmjs.com/package/d3-selection/v/2.0.0) (_removed_)
+  - [codemirror@5.57.0](https://www.npmjs.com/package/codemirror/v/5.57.0)
+  - [d3@6.1.1](https://www.npmjs.com/package/d3/v/6.1.1)
+  - [logrocket@1.0.12](https://www.npmjs.com/package/logrocket/v/1.0.12)
+  - [eslint@7.8.0](https://www.npmjs.com/package/eslint/v/7.8.0)
+  - [js-beautify@1.13.0](https://www.npmjs.com/package/js-beautify/v/1.13.0)
+  - [d3-zoom@2.0.0](https://www.npmjs.com/package/d3-zoom/v/2.0.0)
+  - [sinon@9.0.3](https://www.npmjs.com/package/sinon/v/9.0.3)
+  - [vue-code-highlight@0.7.5](https://www.npmjs.com/package/vue-code-highlight/v/0.7.5)
+  - [vue-router@3.4.3](https://www.npmjs.com/package/vue-router/v/3.4.3)
+  - [start-server-and-test@1.11.3](https://www.npmjs.com/package/start-server-and-test/v/1.11.3)
+  - [stylelint@13.7.0](https://www.npmjs.com/package/stylelint/v/13.7.0)
+  - [vue@2.6.12](https://www.npmjs.com/package/vue/v/2.6.12)
+  - [vue-template-compiler@2.6.12](https://www.npmjs.com/package/vue-template-compiler/v/2.6.12)
+  - [vuetify@2.3.10](https://www.npmjs.com/package/vuetify/v/2.3.10)
+
+### Bugfixes
+
+- Filter scheduled flows from the activity timeline [#188](https://https://github.com/PrefectHQ/ui/pull/188)
+
+## 2020-08-31
+
+### Features and Improvements
+
+- Update scrollbars with platform-specific styling [#177](https://github.com/PrefectHQ/ui/pull/177)
+
+### Bugfixes
+
+- Fix "overenthusiastic" scrollbars on non-Mac systems [#177](https://github.com/PrefectHQ/ui/pull/177)
+- Remove token name length requirement in run flow tutorial [#174](https://github.com/PrefectHQ/ui/pull/174)
+
+## 2020-08-28
+
+### Features and Improvements
+
+- Add unit tests for the refresh and sidenav Vuex stores [#165](https://github.com/PrefectHQ/ui/pull/165)
+- Improve state setting UX and restarting flow runs from failure - [#163](https://github.com/PrefectHQ/ui/pull/163)
+- Remove references to task and flow run durations [#169](https://github.com/PrefectHQ/ui/pull/169)
+
+### Bugfixes
+
+- Fix log generation when restarting flow runs from failure - [#163](https://github.com/PrefectHQ/ui/pull/163)
+- Fix membership invitation toasts [#166](https://github.com/PrefectHQ/ui/pull/166)
+- Fix bad labels references for older flows that were preventing the Flow page details tile from rendering correctly [#171](https://github.com/PrefectHQ/ui/pull/171)
+
+## 2020-08-27
+
+### Features and Improvements
+
+- Remove stale references to `run_count` from all GQL queries - [#157](https://github.com/PrefectHQ/ui/pull/157)
+
+### Bugfixes
+
+- Don't show maintenance mode on load (unless it's true) [#159](https://github.com/PrefectHQ/ui/pull/159)
+
+## 2020-08-26
+
+### Features and Improvements
+
+- Update connection icon if Cloud is in maintenance mode - [#155](https://github.com/PrefectHQ/ui/pull/155)
+
+### Bugfixes
+
+- None
+
 ## 2020-08-25
 
 ### Features and Improvements